harmonic_optimizer/core.py
import numpy as np
from typing import Union, Dict, List, Optional
from dataclasses import dataclass
from scipy import signal

@dataclass
class OptimizationParams:
    '''Parameters for the harmonic optimization algorithm'''
    R: float  # Resonance factor
    F: float  # Fuel efficiency factor
    E: float  # Energy conversion factor
    constant: float  # Optimization constant (phi, pi, e, etc.)

class HarmonicOptimizer:
    '''
    Core implementation of the Harmonic Optimization algorithm.
    
    The algorithm implements the equation: Φ = √(R·F)² + E²
    for signal optimization and noise reduction.
    '''
    
    def __init__(
        self,
        params: OptimizationParams,
        mode: str = 'adaptive',
        tolerance: float = 1e-6
    ):
        self.params = params
        self.mode = mode
        self.tolerance = tolerance
        self._validate_params()
    
    def _validate_params(self) -> None:
        '''Validate initialization parameters'''
        if any(p <= 0 for p in [self.params.R, self.params.F, self.params.E]):
            raise ValueError("All parameters must be positive")
    
    def optimize(
        self,
        signal: np.ndarray,
        noise_profile: Optional[Dict] = None
    ) -> np.ndarray:
        '''
        Optimize the input signal using the harmonic algorithm.
        
        Args:
            signal: Input signal to optimize
            noise_profile: Optional noise characteristics
            
        Returns:
            Optimized signal
        '''
        optimization_factor = self._compute_optimization_factor()
        
        if self.mode == 'adaptive':
            return self._adaptive_optimization(signal, optimization_factor)
        return self._direct_optimization(signal, optimization_factor)
    
    def _compute_optimization_factor(self) -> float:
        '''Compute the core optimization factor'''
        return np.sqrt(self.params.R * self.params.F)**2 + self.params.E**2 * self.params.constant
    
    def _adaptive_optimization(
        self,
        signal: np.ndarray,
        optimization_factor: float
    ) -> np.ndarray:
        '''Implement adaptive optimization based on local signal characteristics'''
        # Estimate local noise level
        local_noise = np.std(signal - signal.medfilt(signal, 5))
        
        # Apply adaptive optimization
        return signal / (1 + optimization_factor * local_noise)
    
    def _direct_optimization(
        self,
        signal: np.ndarray,
        optimization_factor: float
    ) -> np.ndarray:
        '''Implement direct optimization without adaptation'''
        return signal / (1 + optimization_factor)
